+++ Libraries/MINDful/include/simd/intel_sse2/integers/impl/Vector4/Vector4ui/operators.hpp
@@ -18,7 +18,7 @@ static inline Vector4ui operator+(const Vector4ui& a, const Vector4ui& b) {
  * @return  Result of the addition.
  */
 static inline Vector4ui operator+(const Vector4ui& a, const float& b) {
-    return a + Vector4ui(b) ;
+    return a + Vector4ui(static_cast<uint32_t>(b)) ;
 }
 
 /**
@@ -29,7 +29,7 @@ static inline Vector4ui operator+(const Vector4ui& a, const float& b) {
  * @return  Result of the addition.
  */
 static inline Vector4ui operator+(const float& a, const Vector4ui& b) {
-    return Vector4ui(a) + b ;
+    return Vector4ui(static_cast<uint32_t>(a)) + b ;
 }
 
 /**
@@ -52,7 +52,7 @@ static inline Vector4ui& operator+=(Vector4ui& a, const Vector4ui& b) {
  * @return  Result of the addition stored in the first operand.
  */
 static inline Vector4ui& operator+=(Vector4ui& a, const float& b) {
-    a = a + Vector4ui(b) ;
+    a = a + Vector4ui(static_cast<uint32_t>(b)) ;
     return a ;
 }
 
@@ -99,7 +99,7 @@ static inline Vector4ui operator-(const Vector4ui& a, const Vector4ui& b) {
  * @return  Result of the difference.
  */
 static inline Vector4ui operator-(const Vector4ui& a, const float& b) {
-    return a - Vector4ui(b) ;
+    return a - Vector4ui(static_cast<uint32_t>(b)) ;
 }
 
 /**
@@ -110,7 +110,7 @@ static inline Vector4ui operator-(const Vector4ui& a, const float& b) {
  * @return  Result of the difference.
  */
 static inline Vector4ui operator-(const float& a, const Vector4ui& b) {
-    return Vector4ui(a) - b ;
+    return Vector4ui(static_cast<uint32_t>(a)) - b ;
 }
 
 /**
@@ -133,7 +133,7 @@ static inline Vector4ui& operator-=(Vector4ui& a, const Vector4ui& b) {
  * @return  Result of the difference stored in the first operand.
  */
 static inline Vector4ui& operator-=(Vector4ui& a, const float& b) {
-    a = a - Vector4ui(b) ;
+    a = a - Vector4ui(static_cast<uint32_t>(b)) ;
     return a ;
 }
 
@@ -209,7 +209,7 @@ static inline Vector4ui operator*(const Vector4ui& a, const Vector4ui& b) {
  * @return  Result of the product.
  */
 static inline Vector4ui operator*(const Vector4ui& a, const float& b) {
-    return a * Vector4ui(b) ;
+    return a * Vector4ui(static_cast<uint32_t>(b)) ;
 }
 
 /**
@@ -220,7 +220,7 @@ static inline Vector4ui operator*(const Vector4ui& a, const float& b) {
  * @return  Result of the product.
  */
 static inline Vector4ui operator*(const float& a, const Vector4ui& b) {
-    return Vector4ui(a) * b ;
+    return Vector4ui(static_cast<uint32_t>(a)) * b ;
 }
 
 /**
@@ -243,7 +243,7 @@ static inline Vector4ui& operator*=(Vector4ui& a, const Vector4ui& b) {
  * @return  Result of the product stored in the first operand.
  */
 static inline Vector4ui& operator*=(Vector4ui& a, const float& b) {
-    a = a * Vector4ui(b) ;
+    a = a * Vector4ui(static_cast<uint32_t>(b)) ;
     return a ;
 }
